// app/types/job.type.ts
export interface JobRequest {
  id: number;
  title: string;
  employment_type: number;
  minimum_experience: number;
  description: string;
  department: string;
  country: number;
  job_status: number;
  city: string;
  state?: string;
  postal_code: string;
  approved_salary_range_minimum: number;
  approved_salary_range_maximum: number;
  confidential: boolean;
  private: boolean;
  address: number;
  education: number;
  college: number;
  grade_point_average: number;
  cover_letter: number;
  references: number;
  desired_salary: number;
  earliest_start_date: number;
  website: number;
  relocate: number;
  jh_id: number;
  user_id: number;
}